+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `MoveOwnership` to transfer ownership over messaging.
 
+### Changed
+- telemeter: render new counters with `0` value during scraping to avoid [some problems](https://www.section.io/blog/beware-prometheus-counters-that-do-not-begin-at-zero/).
+
 ## [0.1.28] - 2021-10-14
 ### Added
 - Expose `ActorMeta` and `ActorStatusKind`.